+++ lib/verify.js
@@ -7,8 +7,8 @@ http = require("http"),
 https = require("https"),
 url = require("url"),
 jwcrypto = require("jwcrypto"),
-urlparse = require('urlparse'),
 compareAudiences = require('./compare-audiences.js'),
+validation = require('./validation.js'),
 util = require('util'),
 _ = require('underscore');
 
@@ -60,25 +60,9 @@ function extractExtraClaims(claims) {
   return Object.keys(extraClaims).length ? extraClaims : null;
 }
 
-function validateAuthority(authority) {
-  // canary values
-  const scheme = 'https';
-  const directory = '/foo/';
-  const file = 'bar.html';
-  const query = 'q=blah';
-  const anchor = 'content';
-
-  var url = urlparse(scheme + "://" + authority + directory + file + '?' + query + '#' + anchor);
-  if (url.scheme !== scheme || url.authority !== authority ||
-      url.directory !== directory || url.file !== file ||
-      url.query !== query || url.anchor !== anchor) {
-    throw new Error("invalid hostname");
-  }
-}
-
 function extractDomainFromEmail(email) {
   var domain = (/\@(.*)$/).exec(email)[1].toLowerCase();
-  validateAuthority(domain);
+  validation.validateAuthority(domain);
   return domain;
 }
