+++ lib/well-known-parser.js
@@ -4,23 +4,7 @@
 
 var jwcrypto = require("jwcrypto");
 var urlparse = require("urlparse");
-
-function validateAuthority(authority) {
-  // canary values
-  const scheme = 'https';
-  const directory = '/foo/';
-  const file = 'bar.html';
-  const query = 'q=blah';
-  const anchor = 'content';
-
-  var url = urlparse(scheme + "://" + authority + directory + file + '?' + query + '#' + anchor);
-
-  if (url.scheme !== scheme || url.authority !== authority ||
-      url.directory !== directory || url.file !== file ||
-      url.query !== query || url.anchor !== anchor) {
-    throw new Error("invalid hostname (scheme=" + url.scheme + ", authority=" + url.authority + ", directory=" + url.directory + ", file=" + url.file + ", query=" + url.query + ", anchor=" + anchor + ")");
-  }
-}
+var validation = require('./validation.js');
 
 function validateUrlPath(path) {
   if (path[0] !== '/') {
@@ -70,7 +54,7 @@ module.exports = function(doc, allowURLOmission) {
       throw new Error("malformed authority");
     }
     try {
-      validateAuthority(doc.authority);
+      validation.validateAuthority(doc.authority);
     } catch (e) {
       throw new Error("the authority is not a valid hostname");
     }
